+++ CHANGELOG.md
@@ -28,6 +28,7 @@ New features:
 - Add Tilt equalizer biquad combo.
 - Add GraphicEqualizer biquad combo.
 - Support rate adjust for BlachHole on macOS.
+- Add statefile for persisting runtime parameters to file.
 
 Changes:
 - Optimize cpu load in general, and of dithering and delay filters in particular.